% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_exploration.R
\name{dataset_exploration}
\alias{dataset_exploration}
\title{PMxplore shiny application}
\usage{
dataset_exploration(browser = getOption("browser"), ...)
}
\arguments{
\item{...}{arguments passed to \code{shiny::runApp(...)} function.}
}
\value{
A Shiny application
}
\description{
A Shiny application for NONMEM dataset exploration.
}
\examples{
\dontrun{
dataset_exploration()
}
}
