% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_monitor.R
\name{nonmem_monitor}
\alias{nonmem_monitor}
\title{NONMEM Launcher}
\usage{
nonmem_monitor(type = "local", nonmem_exe = NULL, nonmem_call = NULL,
  nmcheck_exe = NULL, nmcheck_call = NULL, nonmem_parafile = NULL,
  execution_dir = "pmxrun", results_subdir = "results", ...)
}
\arguments{
\item{results_subdir}{}

\item{...}{arguments passed to \code{shiny::runApp(...)} function.}
}
\value{
A Shiny application
}
\description{
A Shiny application to monitor and start of NONMEM jobs.
}
\examples{
\dontrun{
nonmem_monitor()
}
}
