% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonmem_toolbox.R
\name{nonmem_toolbox}
\alias{nonmem_toolbox}
\title{NONMEM Toolbox}
\usage{
nonmem_toolbox(
  run_path = NULL,
  nonmem_exe = NULL,
  nonmem_call = NULL,
  nmcheck_exe = NULL,
  nmcheck_call = NULL,
  ...
)
}
\arguments{
\item{run_path}{character string pointing to the path of a run to load at startup.
Optional.}

\item{...}{arguments passed to \code{shiny::runApp(...)} function.}

\item{nm_exe_path}{NONMEM executable path}

\item{nmcheck_exe_path}{NMTRAN check executable path}
}
\value{
A Shiny application
}
\description{
A Shiny application for interactive post-processing of NONMEM run results.
}
\examples{
\dontrun{
nonmem_toolbox()
}
}
