% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverbrowser.R
\name{serverBrowser}
\alias{serverBrowser}
\title{Server Browser Module}
\usage{
serverBrowser(input, output, session, root_directory,
  initial_selection = NULL, file_extensions = NULL,
  folder_shortcuts = NULL, formatting_function = NULL)
}
\arguments{
\item{formatting_function}{}
}
\description{
Server Browser Module
}
