% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverbrowser.R
\name{serverBrowserUI}
\alias{serverBrowserUI}
\title{Server Browser Module UI}
\usage{
serverBrowserUI(id, max_height = "600px")
}
\arguments{
\item{max_height}{}
}
\value{

}
\description{
Server Browser Module UI
}
