% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/serverbrowser.R
\name{serverBrowserUI}
\alias{serverBrowserUI}
\alias{serverbrowser}
\alias{serverBrowser}
\title{Server browser tree module}
\usage{
serverBrowserUI(id, max_height = "500px")

serverBrowser(
  input,
  output,
  session,
  root_directory = "/",
  initial_selection = NULL,
  dir_highlight = NULL,
  file_highlight = NULL,
  folder_shortcuts = NULL
)
}
\arguments{
\item{id}{UI element id}

\item{input}{shiny input object}

\item{output}{shiny output object}

\item{session}{shiny session object}

\item{root_directory}{character. Path of the root directory to browse from.
Defaults to root directory.}

\item{initial_selection}{optional character. Path of the selected folder at
start.}

\item{dir_highlight}{optional character or function. If character, regular
expression matching directory elements to highlight. If function, one
argument function used to determine if the element should be highlighted.}

\item{file_highlight}{optional character or function. If character, regular
expression matching file elements to highlight. If function, one
argument function used to determine if the element should be highlighted.}
}
\value{
A shiny module
}
\description{
Server browser tree module
}
